sql/requirements.txt
flask==2.3.2
requests==2.32.4
supabase==2.7.1
flask-cors==4.0.1
python-dotenv==1.0.1